requirements.txt
flask
numpy
pandas
opencv-python-headless
tensorflow
torch
torchvision
Pillow
requests
openpyxl
